Move Timer to own class
